int -> float

+++ gui/layout.cc
@@ -9,47 +9,47 @@ Layout::Layout() {}
 
 Layout::~Layout() {}
 
-TableLayout::TableLayout(const std::vector<bool> expandable_rows, const std::vector<bool> expandable_cols, int combined_padding)
+TableLayout::TableLayout(const std::vector<bool> expandable_rows, const std::vector<bool> expandable_cols, float combined_padding)
     : expandable_rows_(expandable_rows), expandable_cols_(expandable_cols), combined_padding_(combined_padding) {}
 
-Sizei TableLayout::CalculateMinimumArea(const std::vector<std::shared_ptr<Widget>>& widgets) const {
+Sizef TableLayout::CalculateMinimumArea(const std::vector<std::shared_ptr<Widget>>& widgets) const {
   // todo: include padding
-  std::vector<int> width(expandable_cols_.size(), 0);
-  std::vector<int> height(expandable_rows_.size(), 0);
+  std::vector<float> width(expandable_cols_.size(), 0);
+  std::vector<float> height(expandable_rows_.size(), 0);
 
   for (auto w: widgets) {
     const LayoutData& d = w->Layout();
-    const Sizei& s = w->GetPreferredSize();
+    const Sizef& s = w->GetPreferredSize();
     UpdateMax(&width[d.GetColumn()], s.GetWidth());
     UpdateMax(&height[d.GetRow()], s.GetHeight());
   }
 
-  return Sizei::FromWidthHeight(std::accumulate(width.begin(), width.end(), 0), std::accumulate(height.begin(), height.end(), 0));
+  return Sizef::FromWidthHeight(std::accumulate(width.begin(), width.end(), 0), std::accumulate(height.begin(), height.end(), 0));
 }
 
-void TableLayout::DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const Recti &area) const {
+void TableLayout::DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const Rectf &area) const {
   // todo: include padding
-  std::vector<int> width(expandable_cols_.size(), 0);
-  std::vector<int> height(expandable_rows_.size(), 0);
+  std::vector<float> width(expandable_cols_.size(), 0);
+  std::vector<float> height(expandable_rows_.size(), 0);
 
   for (auto w: *widgets) {
     const LayoutData& d = w->Layout();
-    const Sizei& s = w->GetPreferredSize();
+    const Sizef& s = w->GetPreferredSize();
     UpdateMax(&width[d.GetColumn()], s.GetWidth());
     UpdateMax(&height[d.GetRow()], s.GetHeight());
   }
 
-  const int total_width = std::accumulate(width.begin(), width.end(), 0);
-  const int total_height = std::accumulate(height.begin(), height.end(), 0);
+  const float total_width = std::accumulate(width.begin(), width.end(), 0);
+  const float total_height = std::accumulate(height.begin(), height.end(), 0);
 
-  const int leftover_width = area.GetWidth() - total_width;
-  const int leftover_height = area.GetHeight() - total_height;
+  const float leftover_width = area.GetWidth() - total_width;
+  const float leftover_height = area.GetHeight() - total_height;
 
   const size_t expandable_rows_count = std::count(expandable_rows_.begin(), expandable_rows_.end(), true);
   const size_t expandable_cols_count = std::count(expandable_cols_.begin(), expandable_cols_.end(), true);
 
   if(expandable_rows_count != 0) {
-    const int extra = leftover_height / expandable_rows_count;
+    const float extra = leftover_height / expandable_rows_count;
     for (int i = 0;  i < expandable_rows_.size(); ++ i) {
       if( expandable_rows_[i]) {
         height[i] += extra;
@@ -58,7 +58,7 @@ void TableLayout::DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const 
   }
 
   if(expandable_cols_count != 0) {
-    const int extra = leftover_width / expandable_cols_count;
+    const float extra = leftover_width / expandable_cols_count;
     for (int i = 0;  i < expandable_cols_.size(); ++ i) {
       if( expandable_cols_[i]) {
         width[i] += extra;
@@ -68,8 +68,8 @@ void TableLayout::DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const 
 
   for (auto w: *widgets) {
     const LayoutData& d = w->Layout();
-    int x = area.GetX();
-    int y = area.GetY();
+    float x = area.GetX();
+    float y = area.GetY();
 
     for(int c=0; c<d.GetColumn(); ++c) {
       x += width[c];
@@ -77,15 +77,15 @@ void TableLayout::DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const 
     for(int r=0; r<d.GetRow(); ++r) {
       y += height[r];
     }
-    w->SetRect(Recti::FromTopLeftWidthHeight(y, x, width[d.GetColumn()], height[d.GetRow()]));
+    w->SetRect(Rectf::FromTopLeftWidthHeight(y, x, width[d.GetColumn()], height[d.GetRow()]));
   }
 }
 
-SingleRowLayout::SingleRowLayout(int padding) : padding_(padding) {}
+SingleRowLayout::SingleRowLayout(float padding) : padding_(padding) {}
 
-Sizei SingleRowLayout::CalculateMinimumArea(const std::vector<std::shared_ptr<Widget>> &widgets) const {
-  int width = 0;
-  int height = 0;
+Sizef SingleRowLayout::CalculateMinimumArea(const std::vector<std::shared_ptr<Widget>> &widgets) const {
+  float width = 0;
+  float height = 0;
   bool first = false;
 
   for (auto w: widgets) {
@@ -96,14 +96,14 @@ Sizei SingleRowLayout::CalculateMinimumArea(const std::vector<std::shared_ptr<Wi
     UpdateMax(&height, s.GetHeight());
   }
 
-  return Sizei::FromWidthHeight(width, height);
+  return Sizef::FromWidthHeight(width, height);
 }
 
-void SingleRowLayout::DoLayout(std::vector<std::shared_ptr<Widget>> *widgets, const Recti &area) const {
-  int x = area.GetX();
+void SingleRowLayout::DoLayout(std::vector<std::shared_ptr<Widget>> *widgets, const Rectf &area) const {
+  float x = area.GetX();
   for (auto w: *widgets) {
     const auto& s = w->GetPreferredSize();
-    w->SetRect( Recti::FromTopLeftWidthHeight(area.GetY(), x, s.GetWidth(), s.GetHeight()) );
+    w->SetRect( Rectf::FromTopLeftWidthHeight(area.GetY(), x, s.GetWidth(), s.GetHeight()) );
     x += s.GetWidth() + padding_;
   }
 }
