+++ gui/layout.h
@@ -14,29 +14,29 @@ public:
   Layout();
   virtual ~Layout();
 
-  virtual Sizei CalculateMinimumArea(const std::vector<std::shared_ptr<Widget>>& widgets) const = 0;
-  virtual void DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const Recti& area) const = 0;
+  virtual Sizef CalculateMinimumArea(const std::vector<std::shared_ptr<Widget>>& widgets) const = 0;
+  virtual void DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const Rectf& area) const = 0;
 };
 
 class TableLayout : public Layout {
 public:
-  TableLayout(const std::vector<bool> expandable_rows, const std::vector<bool> expandable_cols, int combined_padding);
-  virtual Sizei CalculateMinimumArea(const std::vector<std::shared_ptr<Widget>>& widgets) const override;
-  virtual void DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const Recti& area) const override;
+  TableLayout(const std::vector<bool> expandable_rows, const std::vector<bool> expandable_cols, float combined_padding);
+  virtual Sizef CalculateMinimumArea(const std::vector<std::shared_ptr<Widget>>& widgets) const override;
+  virtual void DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const Rectf& area) const override;
 
   std::vector<bool> expandable_rows_;
   std::vector<bool> expandable_cols_;
-  int combined_padding_;
+  float combined_padding_;
 };
 
 class SingleRowLayout : public Layout {
 public:
-  SingleRowLayout(int padding);
-  virtual Sizei CalculateMinimumArea(const std::vector<std::shared_ptr<Widget>>& widgets) const override;
-  virtual void DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const Recti& area) const override;
+  SingleRowLayout(float padding);
+  virtual Sizef CalculateMinimumArea(const std::vector<std::shared_ptr<Widget>>& widgets) const override;
+  virtual void DoLayout(std::vector<std::shared_ptr<Widget>>* widgets, const Rectf& area) const override;
 
 private:
-  int padding_;
+  float padding_;
 };
 
 #endif  // GUI_LAYOUT_H
